+++ src/packages/avatarcropper/index.ts
@@ -1,0 +1,7 @@
+import { AvatarCropper } from './avatarcropper'
+
+export type {
+  AvatarCropperProps,
+  AvatarCropperToolbarPosition,
+} from './avatarcropper'
+export default AvatarCropper